% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort_if_file_doesnt_exist.R
\name{abort_if_file_doesnt_exist}
\alias{abort_if_file_doesnt_exist}
\title{Check if a file exists and abort with a message if not}
\usage{
abort_if_file_doesnt_exist(path)
}
\arguments{
\item{path}{A string containing the path to the file to be checked for}
}
\value{
Returns the path (a string) invisibly
}
\description{
Check if a file exists and abort with a message if not
}
