% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio_subfolders.R
\name{create_portfolio_subfolders}
\alias{create_portfolio_subfolders}
\title{A short description of the function}
\usage{
create_portfolio_subfolders(
  portfolio_name_ref_all = NULL,
  project_location = NULL
)
}
\arguments{
\item{portfolio_name_ref_all}{A description of the argument}

\item{project_location}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
