% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{Checks whether a variable is a data frame}
\usage{
data_check(df)
}
\arguments{
\item{df}{A description of the argument}
}
\value{
A logical: TRUE if df is a data frame and has 1 or more rows,
otherwise FALSE
}
\description{
Checks whether a variable is a data frame. Considers also logicals and null values.
}
