% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_audit_file.R
\name{empty_audit_file}
\alias{empty_audit_file}
\title{A short description of the function}
\usage{
empty_audit_file()
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
