% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_emissions_results.R
\name{empty_emissions_results}
\alias{empty_emissions_results}
\title{A short description of the function}
\usage{
empty_emissions_results()
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
