% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_portfolio_overview.R
\name{empty_portfolio_overview}
\alias{empty_portfolio_overview}
\title{A short description of the function}
\usage{
empty_portfolio_overview()
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
