% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_build_version_msg.R
\name{get_build_version_msg}
\alias{get_build_version_msg}
\title{A short description of the function}
\usage{
get_build_version_msg()
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
