% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_user_errors.R
\name{log_user_errors}
\alias{log_user_errors}
\title{Log user errors}
\usage{
log_user_errors(
  error_name,
  suggested_action = "",
  description = "",
  error_file = file.path(.GlobalEnv$outputs_path, .GlobalEnv$portfolio_name_ref_all,
    "report", "user_errors.json"),
  immediate = FALSE
)
}
\arguments{
\item{error_name}{A parameter}

\item{suggested_action}{A parameter}

\item{description}{A parameter}

\item{error_file}{A parameter}

\item{immediate}{A parameter}
}
\value{
Return value
}
\description{
Log user errors
}
