% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit_if_no_pacta_relevant_data.R
\name{quit_if_no_pacta_relevant_data}
\alias{quit_if_no_pacta_relevant_data}
\title{A short description of the function}
\usage{
quit_if_no_pacta_relevant_data(portfolio)
}
\arguments{
\item{portfolio}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
