% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRDS_or_return_alt_data.R
\name{readRDS_or_return_alt_data}
\alias{readRDS_or_return_alt_data}
\title{Return the object stored in a RDS at a specified file path, or an alternate
value/object if the file doesn't exist.}
\usage{
readRDS_or_return_alt_data(filepath, alt_return = NULL)
}
\arguments{
\item{filepath}{a single string containing a file path to a RDS file}

\item{alt_return}{an R object/value to be returned if the file at \code{filepath}
does not exist}
}
\value{
an object, either the object stored in the RDS at the \code{filepath} or
the value of \code{alt_return}
}
\description{
Return the object stored in a RDS at a specified file path, or an alternate
value/object if the file doesn't exist.
}
