% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_if_exists.R
\name{remove_if_exists}
\alias{remove_if_exists}
\title{A short description of the function}
\usage{
remove_if_exists(obj)
}
\arguments{
\item{obj}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
\examples{
obj <- 1L
exists("obj")
remove_if_exists(obj)
exists("obj")

exists("obj2")
remove_if_exists(obj2)

}
