% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_if_exists.R
\name{save_if_exists}
\alias{save_if_exists}
\title{A short description of the function}
\usage{
save_if_exists(df, portfolio_name, save_name, csv_or_rds = "rds")
}
\arguments{
\item{df}{A description of the argument}

\item{portfolio_name}{A description of the argument}

\item{save_name}{A description of the argument}

\item{csv_or_rds}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
