% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_analysis_inputs_path.R
\name{set_analysis_inputs_path}
\alias{set_analysis_inputs_path}
\title{A short description of the function}
\usage{
set_analysis_inputs_path(
  data_location_ext,
  dataprep_ref = .GlobalEnv$datastore_timestamp
)
}
\arguments{
\item{data_location_ext}{A description of the argument}

\item{dataprep_ref}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
