% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_project_parameters.R
\name{set_project_parameters}
\alias{set_project_parameters}
\title{A short description of the function}
\usage{
set_project_parameters(file_path)
}
\arguments{
\item{file_path}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
