% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_webtool_paths.R
\name{set_webtool_paths}
\alias{set_webtool_paths}
\title{A short description of the function}
\usage{
set_webtool_paths(project_root_dir = "working_dir")
}
\arguments{
\item{project_root_dir}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
