% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{A short description of the function}
\usage{
setup_project()
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
