% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_log.R
\name{write_log}
\alias{write_log}
\title{Write log}
\usage{
write_log(msg, file_path = .GlobalEnv$log_path, ...)
}
\arguments{
\item{msg}{A parameter}

\item{file_path}{A parameter}

\item{...}{Additional parameters}
}
\value{
Return value
}
\description{
Write log
}
